Moved columnrange series to subfolder
